Move transforms related to optimizations
