+++ ecmascript/transforms/src/lib.rs
@@ -8,10 +8,7 @@
 #![cfg_attr(test, feature(test))]
 #![recursion_limit = "1024"]
 
-pub use self::{
-    const_modules::const_modules, fixer::fixer, hygiene::hygiene, inline_globals::InlineGlobals,
-    resolver::resolver, simplify::simplifier,
-};
+pub use self::{const_modules::const_modules, fixer::fixer, hygiene::hygiene, resolver::resolver};
 
 #[macro_use]
 mod macros;
@@ -28,7 +25,6 @@ pub mod compat;
 mod const_modules;
 pub mod debug;
 mod fixer;
-mod inline_globals;
 pub mod modules;
 pub mod optimization;
 pub mod pass;
@@ -36,6 +32,5 @@ pub mod proposals;
 pub mod react;
 mod resolver;
 pub mod scope;
-mod simplify;
 pub mod typescript;
 pub mod util;
