+++ ecmascript/transforms/src/optimization/inline_globals.rs
