+++ ecmascript/transforms/src/optimization/simplify.rs
