+++ ecmascript/transforms/src/optimization/simplify/expr.rs
