+++ ecmascript/transforms/src/optimization/simplify/expr/tests.rs
