+++ ecmascript/transforms/src/optimization/simplify/tests.rs
